[#8] Update the leaderboard every 10 minutes
Changes:
- The leaderboard is now updated every 10 minutes

Note:
- This update results in unnecessary recalculations, with all battles being processed every 10 minutes, even if they've already been evaluated. This issue will be addressed in issue #37.
